+++ src/modules/tools/filter/components/SnippetDropdown.vue
@@ -133,6 +133,11 @@ export default {
             required: false,
             default: 0
         },
+        showAllValues: {
+            type: Boolean,
+            required: false,
+            default: false
+        },
         value: {
             type: Array,
             required: false,
@@ -162,7 +167,8 @@ export default {
                 "STARTSWITH",
                 "ENDSWITH"
             ],
-            source: ""
+            source: "",
+            allValues: false
         };
     },
     computed: {
@@ -274,6 +280,20 @@ export default {
             this.addDropdownValueForAdjustment(this.dropdownValue, this.value, adjusting?.adjust?.value, this.delimitor);
 
             if (adjusting?.finish) {
+                if (Array.isArray(this.allValues)) {
+                    if (Array.isArray(adjusting?.adjust?.value)) {
+                        adjusting.adjust.value.forEach(adjustedValue => {
+                            if (!this.allValues.includes(adjustedValue)) {
+                                this.allValues.push(adjustedValue);
+                            }
+                        });
+                    }
+                    this.allValues.forEach(value => {
+                        if (!this.dropdownValue.includes(value)) {
+                            this.dropdownValue.push(value);
+                        }
+                    });
+                }
                 this.setDropdownSelectedAfterAdjustment(this.dropdownValue, this.dropdownSelected, selected => {
                     this.setCurrentSource("adjust");
                     this.dropdownSelected = selected;
@@ -307,7 +327,6 @@ export default {
     },
     mounted () {
         this.initializeIcons();
-
         if (!this.visible) {
             this.dropdownValue = Array.isArray(this.prechecked) ? this.prechecked : [];
             this.dropdownSelected = this.getInitialDropdownSelected(this.prechecked, this.dropdownValue, this.multiselect);
@@ -335,6 +354,9 @@ export default {
                         this.isInitializing = false;
                         this.disable = false;
                         this.emitSnippetPrechecked(this.prechecked, this.snippetId, this.visible);
+                        if (this.showAllValues && this.prechecked === "all") {
+                            this.allValues = this.dropdownSelected;
+                        }
                     });
                 }, error => {
                     this.disable = false;