+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Added
 - New WebGL Render Pipeline
     - core/layers/renderer/webgl: New layer-class methods for rendering WFS, GeoJSON, OAF or VectorBase layers using OL7's WebGL render pipelines
+- Filter:
+    - New parameter `showAllValues` is added. This prevents the values from disappearing when adjusting them.
 
 ### Changed
 - Update GeoJSON, WFS, OAF getFeaturesFilterFunction filtert kann nach exakter Geometrie filtern.